ml/requirements.txt
requests
pillow
pandas
numpy
torch
torchvision
transformers
pillow-avif-plugin
scikit-learn